Ajustes dos jogadores, criação do crud; criação do controler de animais; ajustes da lsitagem dos animais e numeros sorteados

+++ database/migrations/2021_04_01_202019_jogadores.php
@@ -16,8 +16,8 @@ class Jogadores extends Migration
         Schema::create('jogadores', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('nome');
-            $table->integer('telefone');
-            $table->integer('cpf');
+            $table->bigInteger('telefone');
+            $table->bigInteger('cpf');
             $table->timestamps();
         });
     }
